feat: Modify extension to use webpack
