+++ .gitignore
@@ -1,3 +1,6 @@
 sifty.zip
 
 node_modules/*
+
+build/*
+release/*
