+++ extension/test/integration.html
@@ -18,9 +18,5 @@
 </head>
 <body>
     <h3>Integration Tests</h3>
-    <ul>
-      <li><a href="?spec=">See all tests</a></li>
-      <li><a href="?spec=all">Do all tests</a></li>
-    </ul>
 </body>
 </html>
