+++ extension/test/resources/fig.gif
