+++ extension/test/resources/hummingbird.png
