--- src/background.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8">
-    <title id="title">Sifty Background Page</title>
-
-    <script type="module" src="../test/utils/TestHelper.js"></script>
-    <script type="module" src="./sifty.js"></script>
-</head>
-</html>
